+++ lib/config/yml_generator_config.dart
@@ -105,6 +105,7 @@ class YmlGeneratorConfig {
           name: key,
           path: path,
           generateMap: value['generate_map'] == true,
+          generateExtensions: value['generate_extensions'] == true,
           baseDirectory: baseDirectory,
           fields: fields,
           extraImports: extraImports,