+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## [4.3.1] - 2021-xx-xx
 ### Added
 - Added support for generating a mapping for enum's: `generate_map: true`
+- Added support for generating extension methods using the enum's mapping: `generate_extensions: true`
 
 ## [4.3.0] - 2021-04-20
 ### Added